requirements.txt
gpiozero==1.5.1
numpy==1.18.4
opencv-python==4.1.1.26
picamera==1.13
python-telegram-bot==12.7
RPi.GPIO==0.7.0
